+++ CHANGELOG.md
@@ -157,6 +157,7 @@ However, if the shopkeeper actually moved from its previous spawn location and n
 * Debug: The Yaml output of the `/shopkeeper yaml` command is logged as a single multi-line message to the server log now. This should make it easier to copy and paste the output (depends on the logging configuration though). The keys used in the Yaml output are also more context-specific now.
 * Debug: When the result item of a trade does not match the expected result item, we additionally log the serialized Yaml representation of these items now.
 * Fixed: Editing and removing trades from a shopkeeper while another player is trading with it could result in an exception. The issue was that the insertion of empty dummy trades did not properly work.
+* Improved: When trades are removed while a player trades with a shopkeeper, we would previously insert empty dummy trades at the end of the trades list. This ensures that the number of trades never decreases while a player is trading, even if the shopkeeper's trades dynamically change in the meantime. This is required because the Minecraft client does not properly deal with the list of trades dynamically shrinking in size. However, these empty dummy trade are rather unusual and may therefore be confusing to players. Therefore, we now try to (heuristically) guess the trades that were removed and then insert blocked dummy trades that correspond to those.
 
 API:  
 * Renamed TradingOffer to TradeOffer.